+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.0.1 (2026-02-16)
+
+- Add package.json file
+- Add src/app.js file
+- Add src/index.js file
+- Add src/configs/environment.cfg.js file
+- Add src/configs/logger.cfg.js file
+
 ## 1.0.0 (2026-02-15)
 
 - initial release